pyproject.toml
[tool.poetry]
name = "github-task"
version = "0.1.0"
description = ""
authors = ["alex <a556077@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.2.7"
python-dotenv = "^1.1.1"
psycopg2-binary = "^2.9.11"
django-allauth = {extras = ["socialaccount"], version = "^65.12.1"}
d = "^0.2.2"
django-extensions = "^4.1"
flake8 = "^7.3.0"
black = "^25.9.0"
isort = "^7.0.0"
pytest-django = "^4.11.1"
django-debug-toolbar = "^6.0.0"
celery = {extras = ["redis"], version = "^5.5.3"}
django-ratelimit = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"






[tool.black]
line-length = 120
target-version = ["py310"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
default_section = "THIRDPARTY"
force_single_line = true
line_length = 120
use_parentheses = true
py_version = 313
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
ensure_newline_before_comments = true
skip = ["venv", ".venv", "build", "dist", "_build"]
